Refactor of the API folder structure, moved some code and files to a different location. Also moved all V0 code to the V0 folder so that it will be easier to remove it later on
