--- api/app/signals/apps/api/generics/permissions/__init__.py
@@ -1,26 +1,0 @@
-from signals.apps.api.generics.permissions.base import (
-    ModelWritePermissions,
-    SIAPermissions,
-    SignalCreateInitialPermission,
-    SplitPermission
-)
-from signals.apps.api.generics.permissions.v0 import (
-    CategoryPermission,
-    LocationPermission,
-    NotePermission,
-    PriorityPermission,
-    StatusPermission
-)
-
-__all__ = (
-    'StatusPermission',
-    'CategoryPermission',
-    'LocationPermission',
-    'PriorityPermission',
-    'NotePermission',
-
-    'SIAPermissions',
-    'ModelWritePermissions',
-    'SplitPermission',
-    'SignalCreateInitialPermission',
-)
