+++ api/app/signals/apps/api/generics/validators.py
@@ -3,19 +3,6 @@ from rest_framework.serializers import ValidationError
 from signals.apps.signals.models import Signal
 
 
-class NearAmsterdamValidatorMixin:
-
-    def validate_geometrie(self, value):
-        fail_msg = 'Location coordinates not anywhere near Amsterdam. (in WGS84)'
-
-        lat_not_in_adam_area = not 50 < value.coords[1] < 55
-        lon_not_in_adam_area = not 1 < value.coords[0] < 7
-
-        if lon_not_in_adam_area or lat_not_in_adam_area:
-            raise ValidationError(fail_msg)
-        return value
-
-
 class SignalSourceValidator(object):
     requires_context = True
